python/requirements.txt
# =========================================================================================
#  Copyright 2015 Community Information Online Consortium (CIOC) and KCL Software Solutions
#
#  Licensed under the Apache License, Version 2.0 (the "License");
#  you may not use this file except in compliance with the License.
#  You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.
# =========================================================================================

--find-links https://binary-deps.cioc.ca/cioc/
pyodbc==3.0.7
pyramid==1.4.6
Babel==0.9.6
Mako==1.0.0
pyramid_mako==1.0.2
Paste
PasteDeploy
PasteScript
Sphinx==1.0.7
WebHelpers==1.3
backports.ssl-match-hostname==3.2a3
isodate==0.4.4
marrow.interface==1.0.0
marrow.mailer==4.0.0
marrow.util==1.2.2
pyramid-exclog==0.4
pyramid-simpleform==0.6.1
redis==2.4.10
requests==2.5.0
ipaddress==1.0.6
pyramid_redis_sessions==1.0a2
dogpile.core==0.4.1
dogpile.cache==0.5.4
lxml==3.3.5
pycrypto==2.6.1
pyramid_debugtoolbar==2.2
pyramid_multiauth==0.4.0
pyramid_handlers==0.5
